+++ docs/src/releasenotes.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## Unreleased
+* `Serialization.serialize` can use `dill` instead of `pickle` by setting the env var `JULIA_PYTHONCALL_PICKLE=dill`.
+
 ## 0.9.20 (2024-05-01)
 * The IPython extension is now automatically loaded upon import if IPython is detected.
 * JuliaCall now compatible with Julia 1.10.3.